Touchups and Updates
Added a letter and tape to the tutorial. Some folder clean up for the new assets to be favored over the old assets. Updated the scene loader to load the tutorial and go back to the main menu after completing the tutorial. Temporary until The Dead Body puzzle is finished.