Adding Intro Video
This is a little project to try and figure out how to get an intro video to play before the main menu. As of right now, if you just move the intro video scene in front of the main menu, then the play new game button will not work in build.