Updating Action Block
Adding additional basic enemies so that I can begin making changes so that the enemies have to be in an appropriate distance from each other before they can here gun fire. Also, expanded the action block floor plan so there are sections for me to work on the advanced enemy and boss.