Update The Dead Body.unity
Set up the second locked door and cleaned up alittle more. Moved the body to be blocking the door that cant be opened, but can still be squeezed through. Also, added a box to allow the player to return to the previous sections of the level after entering the final lever section.