Renamed The Dead Body & Updated The Build
Changed the name from The Dead Body to The Dead Bodies. Linked up the levers to be activated before the door is unlocked. Minor touch ups to the level that would case the player to be stuck between boxes.

Also, Updated the build so that this level is added and fits in where it should.