Minor bug fixes
Opening tutorial had an issue where the bookcase would freeze gameplay, but that is fixed. The defensive had an issue where if the player had two keys in their possession, then it would consume both opening one door, but that's been fixed. The hounds are free had an issue where the player couldn't advance to the next level, but that's been fixed.