Updating Levels and Build Settings
The Escape and The End is Nigh are prepped for beta touch ups. Added the alternate ending where the player dies after The Escape, but then they load back into The Escape at the beginning, so unfortunately I think their inventory crosses over, but that is a minor issue that could lead to them hoarding items here.